agora-engine/pyproject.toml
[project]
name = "agora-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiomysql>=0.3.2",
    "fastapi[standard]>=0.127.1",
    "pydantic-ai>=1.39.0",
    "python-dotenv>=1.2.1",
]
